safeguard .simplemenu folder cleanup on boot

+++ simplemenu/output/Bittboy/scripts/autoexec.sh
@@ -1,4 +1,7 @@
 #!/bin/sh
-find /mnt/.simplemenu -name '._*' -exec rm -v {} \;
+if [ -d "/mnt/.simplemenu" ]; then
+  echo "Cleaning .simplemenu folder"
+  find /mnt/.simplemenu -name '._*' -exec rm -v {} \;
+fi
 cd /mnt/apps/simplemenu
 ./simplemenu
